[Icon Applet] Repopulate if URL didn't change
We used to always delete the backing .desktop file if we changed a property in a Link (to
a document) hoping that setUrl would call populate. However, it won't do that if the URL
didn't change, so if you only changed the icon, the .desktop file was removed and never
re-created, breaking the applet.

BUG: 380922
FIXED-IN: 5.10.2

Differential Revision: https://phabricator.kde.org/D6148

+++ applets/icon/iconapplet.cpp
@@ -484,12 +484,18 @@ void IconApplet::configure()
     connect(dialog, &KPropertiesDialog::applied, this, [this] {
         KDesktopFile desktopFile(m_localPath);
         if (desktopFile.hasLinkType()) {
-            // make sure to fully repopulate in case the user changed the Link URL
-            QFile::remove(m_localPath);
-            setUrl(QUrl(desktopFile.readUrl())); // calls populate() itself
-        } else {
-            populate();
+            const QUrl newUrl(desktopFile.readUrl());
+
+            if (m_url != newUrl) {
+                // make sure to fully repopulate in case the user changed the Link URL
+                QFile::remove(m_localPath);
+
+                setUrl(newUrl); // calls populate() itself, but only if it changed
+                return;
+            }
         }
+
+        populate();
     });
 
     dialog->setAttribute(Qt::WA_DeleteOnClose, true);
